requirements.txt
streamlit
streamlit-webrtc
streamlit-lottie
streamlit-option-menu
numpy
matplotlib
scipy
pydub
soundfile
sounddevice
librosa
pandas
google-generativeai
supabase
reportlab
fpdf2
plotly
seaborn
Pillow
requests
python-dotenv
pydantic
asyncio
websockets
datetime
uuid
base64
io
threading
time
av
json
